// frontend/src/lib/api/generated/schemas/appealListItemResponseStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Baltiyskiy Bereg Backend API
 * Admin API for the Baltiyskiy Bereg service-desk assistant. Use it for admin authentication, dashboard analytics, appeals management, and assistant settings.
 * OpenAPI spec version: 1.0.0
 */

export type AppealListItemResponseStatus = typeof AppealListItemResponseStatus[keyof typeof AppealListItemResponseStatus];


export const AppealListItemResponseStatus = {
  open: 'open',
  in_progress: 'in_progress',
  closed: 'closed',
} as const;